+++ packages/TokenRunner/src/Wrapper/FixerWrapper/DocBlockWrapper.php
@@ -16,7 +16,7 @@ use phpDocumentor\Reflection\DocBlockFactory;
 use phpDocumentor\Reflection\Types\Array_;
 use phpDocumentor\Reflection\Types\Compound;
 use Symplify\TokenRunner\DocBlock\ArrayResolver;
-use Symplify\TokenRunner\DocBlock\DocBlockSerializerFactory;
+use Symplify\TokenRunner\ReflectionDocBlock\DocBlockSerializerFactory;
 use Symplify\TokenRunner\Guard\TokenTypeGuard;
 
 final class DocBlockWrapper
